--- app/src/main/java/com/example/oopusingkotlin/Dataclass.kt
@@ -1,21 +1,0 @@
-package com.example.oopusingkotlin
-//
-//data class User(val id:Long, var name: String )
-//
-//fun main(){
-//    val user1 = User(1,"mustafa")
-//
-////    val name = user1.name
-//    user1.name = "Osama"
-////    println(name)
-//    val user2 = User(1,"Osama")
-//    println(user1 == user2)
-//    val updatedUser = user1.copy(name = "safeer")
-//    val (id,name) = User(32,"Hamza")
-//    println("$name has id is $id" )
-//    println(user1)
-//    println(user2)
-//    println(updatedUser)
-//    println(updatedUser.component1())
-//    println(updatedUser.component2())
-//}
