+++ app/src/main/java/com/example/oopusingkotlin/Oopusingkotlin.kt
@@ -1,33 +1,57 @@
 package com.example.oopusingkotlin
 
-fun main(){
-//    var mycar = Car()
-//    println(mycar.mybrand)
-//    mycar.maxspeed = 5
-//    println(mycar.maxspeed)
-//    println(mycar.mymodel)
-//    mycar.mymodel = "M2"
-}
+import android.graphics.ColorSpace
 
+fun main(){
+    var mustafa = Person("Mustafa","Anis",32)
+    mustafa.age = 35
+    println("Mustafa is ${mustafa.age} years old")
+    mustafa.stateHobby()
+
+
+
+
+
+//    mustafa.hobby = "Cricket"
+//    var john = Person()
+//    john.hobby = "football"
+//    john.stateHobby()
+//    var johncorner = Person(lastName = "petterson")
+
+}
+class Person(firstName: String = "john", lastName: String = "Doe"){
+    var age : Int? = null
+    var hobby : String = "Watch Netflix"
+    var firstName: String? = null
+    init {
+        this.firstName = firstName
+     println("person created")
+     println("firstname = $firstName lastname = $lastName")
+ }
+    constructor(firstName: String,lastName: String,age: Int)
+            : this(firstName,lastName){
+        this.age = age
+        println("firstname = $firstName lastname = $lastName , age = $age")
+
+    }
+
+    fun stateHobby(){
+        println("$firstName\'s hobby is $hobby")
+    }
+
+
+}
 //
-//class Car(){
-//    lateinit var owner : String
+//fun main(){
+//    var samsung = MobilePhone("android","samsung","A21s")
+//    var iphone  = MobilePhone("ios","iphone","7plus")
+//    var huawei = MobilePhone ("huawei_software","huawei","nova7i")
 //
-//    val mybrand: String = "BMW"
-//    get(){
-//        return field.lowercase()
-//    }
-//    var maxspeed: Int = 2500
-//        set(value) {
-//            field = if (value > 0) value else throw IllegalArgumentException("Max speed cant be less than 0")
-//        }
-//    var mymodel : String = "M3"
-//    private set
-//
+//}
+//class MobilePhone(osName: String,brand: String,model: String){
 //
 //    init {
-//        this.mymodel = "M1"
-//        this.owner = "frank"
+//        println("OsName = $osName brand = $brand model = $model")
 //    }
 //
 //}
